# docs/swagger.yaml
basePath: /
definitions:
  domain.AsyncResponse:
    properties:
      message:
        type: string
      status:
        type: string
      transaction_id:
        type: string
    type: object
  domain.AuthRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  domain.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
  domain.IdentityRecord:
    properties:
      created_at:
        type: string
      demographic_hash:
        type: string
      dob:
        type: string
      gender:
        type: string
      name:
        type: string
      score:
        type: number
      transaction_id:
        type: string
    type: object
  domain.KYCEnrollRequest:
    properties:
      callback_url:
        example: http://client-service.local/webhook
        type: string
      dob:
        example: "1990-01-01"
        type: string
      gender:
        example: FEMALE
        type: string
      name:
        type: string
      photo_base64:
        type: string
    required:
    - dob
    - gender
    - name
    - photo_base64
    type: object
  domain.KYCVerifyRequest:
    properties:
      callback_url:
        example: http://client-service.local/webhook
        type: string
      dob:
        example: "1990-01-01"
        type: string
      gender:
        example: FEMALE
        type: string
      name:
        type: string
      photo_base64:
        type: string
    required:
    - dob
    - gender
    - name
    - photo_base64
    type: object
  domain.TransactionStatus:
    enum:
    - PENDING
    - MATCHED
    - PARTIAL_MATCH
    - NO_MATCH
    - ERROR
    - SUCCESS
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusMatched
    - StatusPartial
    - StatusNoMatch
    - StatusError
    - StatusSuccess
  domain.VerificationDetails:
    properties:
      demographic_match:
        type: boolean
      face_similarity:
        type: number
      name_similarity:
        type: number
    type: object
  domain.VerificationResult:
    properties:
      confidence_score:
        type: number
      created_at:
        type: string
      details:
        $ref: '#/definitions/domain.VerificationDetails'
      status:
        $ref: '#/definitions/domain.TransactionStatus'
      transaction_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: AI-driven KYC system with multi-modal identity signatures.
  title: Distributed KYC Engine API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Exchanges credentials for a JWT access token. (Mock: any username/password
        works)'
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /health:
    get:
      description: Returns the current health status of the service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health Check
      tags:
      - system
  /kyc/enroll:
    post:
      consumes:
      - application/json
      description: Creates a unique identity signature for the user and stores demographic
        metadata.
      parameters:
      - description: Enrollment Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.KYCEnrollRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/domain.AsyncResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: KYC Enrollment
      tags:
      - kyc
  /kyc/search:
    get:
      description: Searches for identity records based on name and/or gender metadata.
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Gender
        in: query
        name: gender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.IdentityRecord'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Search Identities
      tags:
      - kyc
  /kyc/status/{transaction_id}:
    get:
      description: Retrieves the status or result of an asynchronous KYC transaction.
      parameters:
      - description: Transaction ID
        in: path
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VerificationResult'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Transaction Status
      tags:
      - kyc
  /kyc/verify:
    post:
      consumes:
      - application/json
      description: Performs similarity-based lookup for returning users using facial
        embeddings and demographic matching.
      parameters:
      - description: Verification Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.KYCVerifyRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/domain.AsyncResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Re-KYC Verification
      tags:
      - kyc
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"